+++ src/dgisim/character/characters_set.py
@@ -16,6 +16,7 @@ _DEFAULT_CHARACTERS: list[type[Character]] = [
     Diona,
     ElectroHypostasis,
     Eula,
+    FatuiCryoCicinMage,
     FatuiPyroAgent,
     Fischl,
     Ganyu,